+++ R/aaa.R
@@ -1,2 +1,5 @@
 the <- new_environment()
 the$server_processes <- list()
+the$http_allowed_origins <- character()
+the$http_trusted_hosts <- character()
+the$http_shared_secret <- NULL